requirements.txt
psutil~=5.9.7
Flask~=3.0.0
Flask-SQLAlchemy~=3.1.1
Flask-Login~=0.6.3
Werkzeug~=3.0.1
flask-seasurf @ git+https://github.com/maxcountryman/flask-seasurf@f383b48